requirements.txt
huggingface-hub
streamlit
streamlit-webrtc
opencv-contrib-python
mediapipe
numpy
pyyaml
simpleaudio
pydub
python-dotenv           
google-generativeai     
gTTS                   
torch                  
torchvision            
dlib                    
